Counter Component
I believe this is correct, by providing Counters prop object as an argument in the onClick event, our action creator will be able to take account of the current value for count in the store, as (props.count) which we can increment or decrement in the payload. Then in the reducer I will return a new store state that is the result of this change.